refactor: 패키지 구조 변경( 맛집 API 관련 famous_restaurant 패키지에 속하도록 함 )
