+++ app/src/main/java/com/mbj/doeat/data/remote/network/api/famous_restarant/service/SearchService.kt
@@ -1,4 +1,4 @@
-package com.mbj.doeat.data.remote.network.service
+package com.mbj.doeat.data.remote.network.api.famous_restarant.service
 
 import com.mbj.doeat.data.remote.model.SearchResult
 import com.mbj.doeat.data.remote.network.adapter.ApiResponse
